refactor: message::message is now message
