+++ src/message/mod.rs
@@ -2,10 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0, MIT
 
 pub mod chain_message;
-pub mod message;
 pub mod signed_message;
 
 use crate::shim::{address::Address, econ::TokenAmount, message::Message as ShimMessage};
+use crate::shim::{gas::Gas, version::NetworkVersion};
 pub use chain_message::ChainMessage;
 use fvm_ipld_encoding3::RawBytes;
 use fvm_shared::MethodNum;
@@ -90,6 +90,54 @@ impl Message for ShimMessage {
         self.gas_premium = prem.into();
     }
 }
+
+/// Semantic validation and validates the message has enough gas.
+pub fn valid_for_block_inclusion(
+    msg: &ShimMessage,
+    min_gas: Gas,
+    version: NetworkVersion,
+) -> Result<(), anyhow::Error> {
+    use crate::shim::address::ZERO_ADDRESS;
+    use fvm_shared3::{BLOCK_GAS_LIMIT, TOTAL_FILECOIN};
+    if msg.version != 0 {
+        anyhow::bail!("Message version: {} not supported", msg.version);
+    }
+    if msg.to == *ZERO_ADDRESS && version >= NetworkVersion::V7 {
+        anyhow::bail!("invalid 'to' address");
+    }
+    if msg.value.is_negative() {
+        anyhow::bail!("message value cannot be negative");
+    }
+    if msg.value > *TOTAL_FILECOIN {
+        anyhow::bail!("message value cannot be greater than total FIL supply");
+    }
+    if msg.gas_fee_cap.is_negative() {
+        anyhow::bail!("gas_fee_cap cannot be negative");
+    }
+    if msg.gas_premium.is_negative() {
+        anyhow::bail!("gas_premium cannot be negative");
+    }
+    if msg.gas_premium > msg.gas_fee_cap {
+        anyhow::bail!("gas_fee_cap less than gas_premium");
+    }
+    if msg.gas_limit > BLOCK_GAS_LIMIT {
+        anyhow::bail!(
+            "gas_limit {} cannot be greater than block gas limit",
+            msg.gas_limit
+        );
+    }
+
+    if Gas::new(msg.gas_limit) < min_gas {
+        anyhow::bail!(
+            "gas_limit {} cannot be less than cost {} of storing a message on chain",
+            msg.gas_limit,
+            min_gas
+        );
+    }
+
+    Ok(())
+}
+
 #[cfg(test)]
 mod tests {
     mod builder_test;
